chore(node/rpc): Refactor p2p rpc
