+++ crates/node/p2p/Cargo.toml
@@ -19,7 +19,6 @@ workspace = true
 # Kona
 kona-macros.workspace = true
 kona-genesis.workspace = true
-kona-rpc = { workspace = true, features = ["jsonrpsee", "reqwest", "std"] }
 
 # Alloy
 alloy-rlp.workspace = true
@@ -50,10 +49,8 @@ tokio.workspace = true
 tracing.workspace = true
 thiserror.workspace = true
 lazy_static.workspace = true
-async-trait.workspace = true
 unsigned-varint.workspace = true
 rand = { workspace = true, features = ["thread_rng"] }
-jsonrpsee = { workspace = true, features = ["server"] }
 serde_json = { workspace = true, features = ["alloc"] }
 derive_more = { workspace = true, features = ["display", "deref", "debug"] }
