+++ crates/node/rpc/src/p2p.rs
@@ -5,41 +5,24 @@
 //!
 //! [op-node]: https://github.com/ethereum-optimism/optimism/blob/7a6788836984996747193b91901a824c39032bd8/op-node/p2p/rpc_api.go#L45
 
-use crate::NetRpcRequest;
+use alloc::{boxed::Box, string::String, vec::Vec};
 use async_trait::async_trait;
 use jsonrpsee::{
     core::RpcResult,
     types::{ErrorCode, ErrorObject},
 };
-use kona_rpc::{OpP2PApiServer, PeerCount, PeerDump, PeerInfo, PeerStats};
+use kona_p2p::{P2pRpcRequest, PeerCount, PeerDump, PeerInfo, PeerStats};
 use std::net::IpAddr;
 
-/// A type alias for the sender of a [`NetRpcRequest`].
-type RpcReqSender = tokio::sync::mpsc::Sender<NetRpcRequest>;
-
-/// NetworkRpc
-///
-/// This is a server implementation of [`OpP2PApiServer`].
-#[derive(Debug)]
-pub struct NetworkRpc {
-    /// The channel to send [`NetRpcRequest`]s.
-    pub sender: RpcReqSender,
-}
-
-impl NetworkRpc {
-    /// Constructs a new [`NetworkRpc`] given a sender channel.
-    pub const fn new(sender: RpcReqSender) -> Self {
-        Self { sender }
-    }
-}
+use crate::{OpP2PApiServer, net::NetworkRpc};
 
 #[async_trait]
 impl OpP2PApiServer for NetworkRpc {
     async fn opp2p_self(&self) -> RpcResult<PeerInfo> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_self");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_self");
         let (tx, rx) = tokio::sync::oneshot::channel();
         self.sender
-            .send(NetRpcRequest::PeerInfo(tx))
+            .send(P2pRpcRequest::PeerInfo(tx))
             .await
             .map_err(|_| ErrorObject::from(ErrorCode::InternalError))?;
 
@@ -47,10 +30,10 @@ impl OpP2PApiServer for NetworkRpc {
     }
 
     async fn opp2p_peer_count(&self) -> RpcResult<PeerCount> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_peerCount");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_peerCount");
         let (tx, rx) = tokio::sync::oneshot::channel();
         self.sender
-            .send(NetRpcRequest::PeerCount(tx))
+            .send(P2pRpcRequest::PeerCount(tx))
             .await
             .map_err(|_| ErrorObject::from(ErrorCode::InternalError))?;
 
@@ -61,22 +44,22 @@ impl OpP2PApiServer for NetworkRpc {
     }
 
     async fn opp2p_peers(&self) -> RpcResult<PeerDump> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_peers");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_peers");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_peer_stats(&self) -> RpcResult<PeerStats> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_peerStats");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_peerStats");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_discovery_table(&self) -> RpcResult<Vec<String>> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_discoveryTable");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_discoveryTable");
         let (tx, rx) = tokio::sync::oneshot::channel();
         self.sender
-            .send(NetRpcRequest::DiscoveryTable(tx))
+            .send(P2pRpcRequest::DiscoveryTable(tx))
             .await
             .map_err(|_| ErrorObject::from(ErrorCode::InternalError))?;
 
@@ -84,73 +67,78 @@ impl OpP2PApiServer for NetworkRpc {
     }
 
     async fn opp2p_block_peer(&self, _peer: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_blockPeer");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_blockPeer");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_list_blocked_peers(&self) -> RpcResult<Vec<String>> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_listBlockedPeers");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_listBlockedPeers");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_block_addr(&self, _ip: IpAddr) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_blockAddr");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_blockAddr");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_unblock_addr(&self, _ip: IpAddr) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_unblockAddr");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_unblockAddr");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_list_blocked_addrs(&self) -> RpcResult<Vec<IpAddr>> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_listBlockedAddrs");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_listBlockedAddrs");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_block_subnet(&self, _subnet: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_blockSubnet");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_blockSubnet");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_unblock_subnet(&self, _subnet: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_unblockSubnet");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_unblockSubnet");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_list_blocked_subnets(&self) -> RpcResult<Vec<String>> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_listBlockedSubnets");
+        kona_macros::inc!(
+            gauge,
+            kona_p2p::Metrics::RPC_CALLS,
+            "method",
+            "opp2p_listBlockedSubnets"
+        );
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_protect_peer(&self, _peer: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_protectPeer");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_protectPeer");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_unprotect_peer(&self, _peer: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_unprotectPeer");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_unprotectPeer");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_connect_peer(&self, _peer: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_connectPeer");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_connectPeer");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
 
     async fn opp2p_disconnect_peer(&self, _peer: String) -> RpcResult<()> {
-        kona_macros::inc!(gauge, crate::Metrics::RPC_CALLS, "method", "opp2p_disconnectPeer");
+        kona_macros::inc!(gauge, kona_p2p::Metrics::RPC_CALLS, "method", "opp2p_disconnectPeer");
         // Method not supported yet.
         Err(ErrorObject::from(ErrorCode::MethodNotFound))
     }
