+++ crates/node/service/src/service/validator.rs
@@ -9,9 +9,9 @@ use async_trait::async_trait;
 use kona_derive::traits::{Pipeline, SignalReceiver};
 use kona_engine::EngineStateBuilderError;
 use kona_genesis::RollupConfig;
-use kona_p2p::{Network, NetworkRpc};
+use kona_p2p::Network;
 use kona_protocol::BlockInfo;
-use kona_rpc::{OpP2PApiServer, RpcLauncher, RpcLauncherError};
+use kona_rpc::{NetworkRpc, OpP2PApiServer, RpcLauncher, RpcLauncherError};
 use std::fmt::Display;
 use tokio::sync::mpsc::{self, UnboundedSender};
 use tokio_util::sync::CancellationToken;
